Improved text to reduce interpretation ambiguity
The original text 'Some C API examples are included in the app folder'
is misleading and is implying to the reader to look/search for examples
directly under axis repo 'web-server/app' directory itself and complain
as examples missing. But in reality, those examples are available (as
part of build procedure coded in Dockerfile) inside the container and
hence suggested text is trying to address that concern.

Closes AxisCommunications/acap-native-sdk-examples#87