fix(object-detection-cv25): Fix broken link in README
This reverts the link target to what it was before it was changed in 730ed81fe3bedbf5c92a9e7dc81e2dc52a4310d5. `vdo-larod-cv25` does not and has never existed in the repository and, since the context of the link did not change, I assume it should still point to the same example as it used to.